requirements.txt
streamlit>=1.31.0
opencv-python>=4.8.0
numpy>=1.24.0
torch>=2.0.0
torchvision>=0.15.0
segmentation-models-pytorch>=0.3.3
albumentations>=1.3.1
scikit-image>=0.21.0
pillow>=10.0.0
matplotlib>=3.7.0
plotly>=5.14.0
seaborn>=0.12.0
tensorboard>=2.13.0
transformers>=4.35.0
python-dotenv>=1.0.0
pandas>=2.0.0
